backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.14.0
psycopg2-binary==2.9.9
greenlet==3.1.1

# Redis
redis==5.2.0
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.6.0

# HTTP client
httpx==0.27.2

# Utilities
python-dateutil==2.9.0

# Development/Linting
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pylint==3.3.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
